# Improve performance of creating hasher and signer