--- a/hasher/Cargo.toml
+++ b/hasher/Cargo.toml
@@ -19,3 +19,10 @@ serde = { version = "1.0", features = ["derive"] }
 
 [dev-dependencies]
 serde_json = { version = "1.0" }
+
+# benchmarks
+criterion = "0.3"
+
+[[bench]]
+name = "pasta_criterion"
+harness = false